refactor: move plugin tests to internal
Match the structure of the `./src` folder in `./tests` for the internal
plugins tests.
