+++ v-next/core/test/internal/plugins/fixture-projects/installed-package/README.md
