+++ v-next/core/test/internal/plugins/fixture-projects/installed-peer-deps-as-sub-node-modules/README.md
