+++ v-next/core/test/internal/plugins/fixture-projects/installed-peer-deps/README.md
