+++ v-next/core/test/internal/plugins/fixture-projects/not-installed-package/README.md
