+++ v-next/core/test/internal/plugins/fixture-projects/not-installed-peer-dep/README.md
