+++ v-next/core/test/internal/plugins/fixture-projects/peer-dep-with-wrong-version/README.md
